bzl: Remove library target from Rust crate
