+++ docker-images/syntax-highlighter/crates/scip-treesitter-cli/BUILD.bazel
@@ -1,26 +1,6 @@
 load("@crate_index//:defs.bzl", "aliases", "all_crate_deps")
 load("@rules_rust//rust:defs.bzl", "rust_binary", "rust_library", "rust_test")
 
-rust_library(
-    name = "scip-treesitter-cli-lib",
-    srcs = glob([
-        "src/*.rs",
-    ]),
-    aliases = aliases(),
-    crate_name = "scip_treesitter_cli",
-    crate_root = "src/lib.rs",
-    proc_macro_deps = all_crate_deps(
-        proc_macro = True,
-    ),
-    visibility = ["//visibility:public"],
-    deps = all_crate_deps(
-        normal = True,
-    ) + [
-        "//docker-images/syntax-highlighter/crates/scip-syntax",
-        "//docker-images/syntax-highlighter/crates/scip-treesitter",
-        "//docker-images/syntax-highlighter/crates/scip-treesitter-languages",
-    ],
-)
 
 rust_binary(
     name = "scip-treesitter-cli",
@@ -33,7 +13,6 @@ rust_binary(
     deps = all_crate_deps(
         normal = True,
     ) + [
-        ":scip-treesitter-cli-lib",
         "//docker-images/syntax-highlighter/crates/scip-syntax",
         "//docker-images/syntax-highlighter/crates/scip-treesitter",
         "//docker-images/syntax-highlighter/crates/scip-treesitter-languages",
